// final project - car rent/CarRentWebsite-FinalPro/src/app/models/carAndType.ts
export class CarAndType{
    CarID: number;
    CurrentMileage: number;
    GoodForRent: boolean;
    CarNumber: string;
    Branch: number;
    CarImage: string;
    CarTypeId: number;
    SupplierName: string;
    CarModel: string;
    DailyCost: number;
    DailyDelayCost: number;
    ProductionYear: number;
    Transmission: boolean;
    
}